requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2025.6
distlib==0.3.9
Django==5.1.3
django-admin-interface==0.29.2
django-colorfield==0.11.0
django-crispy-forms==2.4
django-filter==24.3
django-sendgrid-v5==1.2.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
filelock==3.16.1
idna==3.10
mysql-connector-python==9.1.0
mysqlclient==2.2.6
pillow==11.0.0
platformdirs==4.3.6
psycopg==3.2.6
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
python-http-client==3.3.7
python-slugify==8.0.4
requests==2.32.3
sendgrid==6.11.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
text-unidecode==1.3
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.27.1
